Moved User in node-api into internal package.
This class is used in node configuration and by the driver.
