+++ client/rpc/src/test/kotlin/net/corda/client/rpc/RPCPermissionsTests.kt
@@ -4,7 +4,7 @@ import net.corda.core.messaging.RPCOps
 import net.corda.node.services.messaging.getRpcContext
 import net.corda.node.services.messaging.requirePermission
 import net.corda.nodeapi.PermissionException
-import net.corda.nodeapi.User
+import net.corda.nodeapi.internal.User
 import net.corda.testing.RPCDriverExposedDSLInterface
 import net.corda.testing.rpcDriver
 import org.junit.Test
