+++ node/src/main/kotlin/net/corda/node/shell/InteractiveShell.kt
@@ -30,7 +30,7 @@ import net.corda.node.services.messaging.RpcContext
 import net.corda.node.services.statemachine.FlowStateMachineImpl
 import net.corda.node.utilities.ANSIProgressRenderer
 import net.corda.nodeapi.ArtemisMessagingComponent
-import net.corda.nodeapi.User
+import net.corda.nodeapi.internal.User
 import org.crsh.command.InvocationContext
 import org.crsh.console.jline.JLineProcessor
 import org.crsh.console.jline.TerminalFactory
