+++ node/src/smoke-test/kotlin/net/corda/node/CordappSmokeTest.kt
@@ -11,7 +11,7 @@ import net.corda.core.internal.list
 import net.corda.core.messaging.startFlow
 import net.corda.core.utilities.getOrThrow
 import net.corda.core.utilities.unwrap
-import net.corda.nodeapi.User
+import net.corda.nodeapi.internal.User
 import net.corda.smoketesting.NodeConfig
 import net.corda.smoketesting.NodeProcess
 import org.assertj.core.api.Assertions.assertThat
