+++ node/src/test/kotlin/net/corda/node/CordaRPCOpsImplTest.kt
@@ -31,7 +31,7 @@ import net.corda.node.services.FlowPermissions.Companion.startFlowPermission
 import net.corda.nodeapi.internal.ServiceInfo
 import net.corda.node.services.transactions.SimpleNotaryService
 import net.corda.nodeapi.PermissionException
-import net.corda.nodeapi.User
+import net.corda.nodeapi.internal.User
 import net.corda.testing.chooseIdentity
 import net.corda.testing.expect
 import net.corda.testing.expectEvents
