+++ node/src/test/kotlin/net/corda/node/services/config/FullNodeConfigurationTest.kt
@@ -1,7 +1,7 @@
 package net.corda.node.services.config
 
 import net.corda.core.utilities.NetworkHostAndPort
-import net.corda.nodeapi.User
+import net.corda.nodeapi.internal.User
 import net.corda.testing.ALICE
 import net.corda.testing.node.MockServices.Companion.makeTestDataSourceProperties
 import net.corda.testing.node.MockServices.Companion.makeTestDatabaseProperties
