+++ samples/attachment-demo/src/main/kotlin/net/corda/attachmentdemo/Main.kt
@@ -6,7 +6,7 @@ import net.corda.testing.DUMMY_BANK_A
 import net.corda.testing.DUMMY_BANK_B
 import net.corda.testing.DUMMY_NOTARY
 import net.corda.node.services.transactions.SimpleNotaryService
-import net.corda.nodeapi.User
+import net.corda.nodeapi.internal.User
 import net.corda.testing.driver.driver
 
 /**
