+++ testing/node-driver/src/integration-test/kotlin/net/corda/testing/FlowStackSnapshotTest.kt
@@ -12,7 +12,7 @@ import net.corda.core.internal.read
 import net.corda.core.messaging.startFlow
 import net.corda.core.serialization.CordaSerializable
 import net.corda.node.services.FlowPermissions.Companion.startFlowPermission
-import net.corda.nodeapi.User
+import net.corda.nodeapi.internal.User
 import net.corda.testing.driver.driver
 import org.junit.Ignore
 import org.junit.Test
