+++ testing/node-driver/src/main/kotlin/net/corda/testing/DriverConstants.kt
@@ -7,7 +7,7 @@ import net.corda.core.internal.concurrent.transpose
 import net.corda.core.messaging.CordaRPCOps
 import net.corda.nodeapi.internal.ServiceInfo
 import net.corda.node.services.transactions.ValidatingNotaryService
-import net.corda.nodeapi.User
+import net.corda.nodeapi.internal.User
 import net.corda.testing.driver.DriverDSLExposedInterface
 
 //
