+++ testing/node-driver/src/main/kotlin/net/corda/testing/node/NodeBasedTest.kt
@@ -17,7 +17,7 @@ import net.corda.node.services.config.plus
 import net.corda.node.services.network.NetworkMapService
 import net.corda.node.services.transactions.RaftValidatingNotaryService
 import net.corda.node.utilities.ServiceIdentityGenerator
-import net.corda.nodeapi.User
+import net.corda.nodeapi.internal.User
 import net.corda.nodeapi.config.parseAs
 import net.corda.testing.DUMMY_MAP
 import net.corda.testing.TestDependencyInjectionBase
