+++ testing/smoke-test-utils/src/main/kotlin/net/corda/smoketesting/NodeConfig.kt
@@ -6,7 +6,7 @@ import com.typesafe.config.ConfigRenderOptions
 import com.typesafe.config.ConfigValue
 import com.typesafe.config.ConfigValueFactory
 import net.corda.core.identity.CordaX500Name
-import net.corda.nodeapi.User
+import net.corda.nodeapi.internal.User
 
 class NodeConfig(
         val legalName: CordaX500Name,
