+++ tools/demobench/src/main/kotlin/net/corda/demobench/model/NodeConfig.kt
@@ -2,7 +2,7 @@ package net.corda.demobench.model
 
 import com.typesafe.config.*
 import net.corda.core.identity.CordaX500Name
-import net.corda.nodeapi.User
+import net.corda.nodeapi.internal.User
 import java.io.File
 import java.nio.file.Files
 import java.nio.file.Path
