+++ tools/loadtest/src/main/kotlin/net/corda/loadtest/LoadTestConfiguration.kt
@@ -1,6 +1,6 @@
 package net.corda.loadtest
 
-import net.corda.nodeapi.User
+import net.corda.nodeapi.internal.User
 import java.nio.file.Path
 import java.util.concurrent.ForkJoinPool
